bot/requirements.txt
# ═══════════════════════════════════════════════════════════════════════════
#                        NEXUS AI - SOLANA EDITION
#                            Dependencies
# ═══════════════════════════════════════════════════════════════════════════

# ─────────────────────────────────────────────────────────────────────────────
#                              CORE
# ─────────────────────────────────────────────────────────────────────────────
python-dotenv==1.0.0      # Environment variables
loguru==0.7.2             # Logging
pydantic==2.5.0           # Data validation

# ─────────────────────────────────────────────────────────────────────────────
#                              AI
# ─────────────────────────────────────────────────────────────────────────────
openai==1.12.0            # DeepSeek API (OpenAI compatible)

# ─────────────────────────────────────────────────────────────────────────────
#                           SOLANA
# ─────────────────────────────────────────────────────────────────────────────
solana==0.32.0            # Solana SDK
solders==0.21.0           # Solana types
anchorpy==0.20.1          # Anchor framework
base58==2.1.1             # Base58 encoding

# ─────────────────────────────────────────────────────────────────────────────
#                           HTTP / ASYNC
# ─────────────────────────────────────────────────────────────────────────────
httpx==0.27.0             # Async HTTP client
aiohttp==3.9.0            # Async HTTP
websockets==12.0          # WebSocket connections

# ─────────────────────────────────────────────────────────────────────────────
#                           TELEGRAM
# ─────────────────────────────────────────────────────────────────────────────
python-telegram-bot==21.0  # Telegram Bot API

# ─────────────────────────────────────────────────────────────────────────────
#                           DATA
# ─────────────────────────────────────────────────────────────────────────────
pandas==2.1.0             # Data analysis
numpy==1.26.0             # Numerical computing

# ─────────────────────────────────────────────────────────────────────────────
#                           SCHEDULING
# ─────────────────────────────────────────────────────────────────────────────
apscheduler==3.10.4       # Job scheduling

# ─────────────────────────────────────────────────────────────────────────────
#                           UTILITIES
# ─────────────────────────────────────────────────────────────────────────────
ratelimit==2.2.1          # Rate limiting
tenacity==8.2.3           # Retry logic
cachetools==5.3.2         # Caching

# ─────────────────────────────────────────────────────────────────────────────
#                      INFRASTRUCTURE (New)
# ─────────────────────────────────────────────────────────────────────────────
asyncio-throttle==1.0.2   # Async rate limiting
aiofiles==23.2.1          # Async file operations

# ─────────────────────────────────────────────────────────────────────────────
#                      TESTING & BENCHMARKING
# ─────────────────────────────────────────────────────────────────────────────
pytest==8.0.0             # Testing framework
pytest-asyncio==0.23.0    # Async test support
pytest-cov==4.1.0         # Coverage
